Backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-extensions==3.2.3
pydot==1.4.2
pyparsing==3.1.0
sqlparse==0.4.4
Pillow==9.5.0
djangorestframework
django-cors-headers
coreapi==2.3.3
